Hemos llevado a cabo la funcionalida del mapa en su complejidad. Hemos realizado que tras pulsar en un marcador se trasladará la información a la pantalla "información". closes #17, closes #18, closes #19

+++ app/src/main/java/com/example/practicatema7pmdm/DataBaseManager/Esquema.java
@@ -18,8 +18,8 @@ public class Esquema {
 
         public static final String COLUMN_TYPE_ID = "INTEGER";
         public static final String COLUMN_TYPE_NOMBRE = "TEXT";
-        public static final String COLUMN_TYPE_LATITUD = "FLOAT";
-        public static final String COLUMN_TYPE_LONGITUD = "FLOAT";
+        public static final String COLUMN_TYPE_LATITUD = "DOUBLE";
+        public static final String COLUMN_TYPE_LONGITUD = "DOUBLE";
         public static final String COLUMN_TYPE_COMENTARIOS = "TEXT";
         public static final String COLUMN_TYPE_VALORACION = "FLOAT";
         public static final String COLUMN_TYPE_CATEGORIA = "INTEGER";
